requirements.txt
websocket-client
websocket-server
clipboard
pyperclip